Enable TRNG int only when reading random number
Leaving the TRNG interrupt enabled at all times results in 1.4M TRNG interrupts/sec that significantly slow other processing. This version enables the TRNG interrupt when a random number is requested, then immediately disables it after the number is read. Surprisingly, it's slightly faster than the previous version; getting a random number with the new version takes an average of 700.6528 ns vs 701.341 ns with the old.